[Refactor] only for style-transfer
