--- cs231n/.gitignore
@@ -1,3 +1,0 @@
-build/*
-im2col_cython.c
-im2col_cython.so
