--- cs231n/layers.py
@@ -1,301 +1,0 @@
-import numpy as np
-
-
-def affine_forward(x, w, b):
-    """
-    Computes the forward pass for an affine (fully-connected) layer.
-
-    The input x has shape (N, d_1, ..., d_k) where x[i] is the ith input.
-    We multiply this against a weight matrix of shape (D, M) where
-    D = \prod_i d_i
-
-    Inputs:
-    x - Input data, of shape (N, d_1, ..., d_k)
-    w - Weights, of shape (D, M)
-    b - Biases, of shape (M,)
-
-    Returns a tuple of:
-    - out: output, of shape (N, M)
-    - cache: (x, w, b)
-    """
-    out = x.reshape(x.shape[0], -1).dot(w) + b
-    cache = (x, w, b)
-    return out, cache
-
-
-def affine_backward(dout, cache):
-    """
-    Computes the backward pass for an affine layer.
-
-    Inputs:
-    - dout: Upstream derivative, of shape (N, M)
-    - cache: Tuple of:
-      - x: Input data, of shape (N, d_1, ... d_k)
-      - w: Weights, of shape (D, M)
-
-    Returns a tuple of:
-    - dx: Gradient with respect to x, of shape (N, d1, ..., d_k)
-    - dw: Gradient with respect to w, of shape (D, M)
-    - db: Gradient with respect to b, of shape (M,)
-    """
-    x, w, b = cache
-    dx = dout.dot(w.T).reshape(x.shape)
-    dw = x.reshape(x.shape[0], -1).T.dot(dout)
-    db = np.sum(dout, axis=0)
-    return dx, dw, db
-
-
-def relu_forward(x):
-    """
-    Computes the forward pass for a layer of rectified linear units (ReLUs).
-
-    Input:
-    - x: Inputs, of any shape
-
-    Returns a tuple of:
-    - out: Output, of the same shape as x
-    - cache: x
-    """
-    out = np.maximum(0, x)
-    cache = x
-    return out, cache
-
-
-def relu_backward(dout, cache):
-    """
-    Computes the backward pass for a layer of rectified linear units (ReLUs).
-
-    Input:
-    - dout: Upstream derivatives, of any shape
-    - cache: Input x, of same shape as dout
-
-    Returns:
-    - dx: Gradient with respect to x
-    """
-    x = cache
-    dx = np.where(x > 0, dout, 0)
-    return dx
-
-
-def batchnorm_forward(x, gamma, beta, bn_param):
-    """
-    Forward pass for batch normalization.
-
-    During training the sample mean and (uncorrected) sample variance are
-    computed from minibatch statistics and used to normalize the incoming data.
-    During training we also keep an exponentially decaying running mean of the mean
-    and variance of each feature, and these averages are used to normalize data
-    at test-time.
-
-    At each timestep we update the running averages for mean and variance using
-    an exponential decay based on the momentum parameter:
-
-    running_mean = momentum * running_mean + (1 - momentum) * sample_mean
-    running_var = momentum * running_var + (1 - momentum) * sample_var
-
-    Note that the batch normalization paper suggests a different test-time
-    behavior: they compute sample mean and variance for each feature using a
-    large number of training images rather than using a running average. For
-    this implementation we have chosen to use running averages instead since
-    they do not require an additional estimation step; the torch7 implementation
-    of batch normalization also uses running averages.
-
-    Input:
-    - x: Data of shape (N, D)
-    - gamma: Scale parameter of shape (D,)
-    - beta: Shift paremeter of shape (D,)
-    - bn_param: Dictionary with the following keys:
-      - mode: 'train' or 'test'; required
-      - eps: Constant for numeric stability
-      - momentum: Constant for running mean / variance.
-      - running_mean: Array of shape (D,) giving running mean of features
-      - running_var Array of shape (D,) giving running variance of features
-
-    Returns a tuple of:
-    - out: of shape (N, D)
-    - cache: A tuple of values needed in the backward pass
-    """
-    mode = bn_param['mode']
-    eps = bn_param.get('eps', 1e-5)
-    momentum = bn_param.get('momentum', 0.9)
-
-    N, D = x.shape
-    running_mean = bn_param.get('running_mean', np.zeros(D, dtype=x.dtype))
-    running_var = bn_param.get('running_var', np.zeros(D, dtype=x.dtype))
-
-    out, cache = None, None
-    if mode == 'train':
-        # Compute output
-        mu = x.mean(axis=0)
-        xc = x - mu
-        var = np.mean(xc ** 2, axis=0)
-        std = np.sqrt(var + eps)
-        xn = xc / std
-        out = gamma * xn + beta
-
-        cache = (mode, x, gamma, xc, std, xn, out)
-
-        # Update running average of mean
-        running_mean *= momentum
-        running_mean += (1 - momentum) * mu
-
-        # Update running average of variance
-        running_var *= momentum
-        running_var += (1 - momentum) * var
-    elif mode == 'test':
-        # Using running mean and variance to normalize
-        std = np.sqrt(running_var + eps)
-        xn = (x - running_mean) / std
-        out = gamma * xn + beta
-        cache = (mode, x, xn, gamma, beta, std)
-    else:
-        raise ValueError('Invalid forward batchnorm mode "%s"' % mode)
-
-    # Store the updated running means back into bn_param
-    bn_param['running_mean'] = running_mean
-    bn_param['running_var'] = running_var
-
-    return out, cache
-
-
-def batchnorm_backward(dout, cache):
-    """
-    Backward pass for batch normalization.
-
-    For this implementation, you should write out a computation graph for
-    batch normalization on paper and propagate gradients backward through
-    intermediate nodes.
-
-    Inputs:
-    - dout: Upstream derivatives, of shape (N, D)
-    - cache: Variable of intermediates from batchnorm_forward.
-
-    Returns a tuple of:
-    - dx: Gradient with respect to inputs x, of shape (N, D)
-    - dgamma: Gradient with respect to scale parameter gamma, of shape (D,)
-    - dbeta: Gradient with respect to shift parameter beta, of shape (D,)
-    """
-    mode = cache[0]
-    if mode == 'train':
-        mode, x, gamma, xc, std, xn, out = cache
-
-        N = x.shape[0]
-        dbeta = dout.sum(axis=0)
-        dgamma = np.sum(xn * dout, axis=0)
-        dxn = gamma * dout
-        dxc = dxn / std
-        dstd = -np.sum((dxn * xc) / (std * std), axis=0)
-        dvar = 0.5 * dstd / std
-        dxc += (2.0 / N) * xc * dvar
-        dmu = np.sum(dxc, axis=0)
-        dx = dxc - dmu / N
-    elif mode == 'test':
-        mode, x, xn, gamma, beta, std = cache
-        dbeta = dout.sum(axis=0)
-        dgamma = np.sum(xn * dout, axis=0)
-        dxn = gamma * dout
-        dx = dxn / std
-    else:
-        raise ValueError(mode)
-
-    return dx, dgamma, dbeta
-
-
-def spatial_batchnorm_forward(x, gamma, beta, bn_param):
-    """
-    Computes the forward pass for spatial batch normalization.
-
-    Inputs:
-    - x: Input data of shape (N, C, H, W)
-    - gamma: Scale parameter, of shape (C,)
-    - beta: Shift parameter, of shape (C,)
-    - bn_param: Dictionary with the following keys:
-      - mode: 'train' or 'test'; required
-      - eps: Constant for numeric stability
-      - momentum: Constant for running mean / variance. momentum=0 means that
-        old information is discarded completely at every time step, while
-        momentum=1 means that new information is never incorporated. The
-        default of momentum=0.9 should work well in most situations.
-      - running_mean: Array of shape (D,) giving running mean of features
-      - running_var Array of shape (D,) giving running variance of features
-
-    Returns a tuple of:
-    - out: Output data, of shape (N, C, H, W)
-    - cache: Values needed for the backward pass
-    """
-    N, C, H, W = x.shape
-    x_flat = x.transpose(0, 2, 3, 1).reshape(-1, C)
-    out_flat, cache = batchnorm_forward(x_flat, gamma, beta, bn_param)
-    out = out_flat.reshape(N, H, W, C).transpose(0, 3, 1, 2)
-    return out, cache
-
-
-def spatial_batchnorm_backward(dout, cache):
-    """
-    Computes the backward pass for spatial batch normalization.
-
-    Inputs:
-    - dout: Upstream derivatives, of shape (N, C, H, W)
-    - cache: Values from the forward pass
-
-    Returns a tuple of:
-    - dx: Gradient with respect to inputs, of shape (N, C, H, W)
-    - dgamma: Gradient with respect to scale parameter, of shape (C,)
-    - dbeta: Gradient with respect to shift parameter, of shape (C,)
-    """
-    N, C, H, W = dout.shape
-    dout_flat = dout.transpose(0, 2, 3, 1).reshape(-1, C)
-    dx_flat, dgamma, dbeta = batchnorm_backward(dout_flat, cache)
-    dx = dx_flat.reshape(N, H, W, C).transpose(0, 3, 1, 2)
-    return dx, dgamma, dbeta
-
-
-def svm_loss(x, y):
-    """
-    Computes the loss and gradient using for multiclass SVM classification.
-
-    Inputs:
-    - x: Input data, of shape (N, C) where x[i, j] is the score for the jth class
-      for the ith input.
-    - y: Vector of labels, of shape (N,) where y[i] is the label for x[i] and
-      0 <= y[i] < C
-
-    Returns a tuple of:
-    - loss: Scalar giving the loss
-    - dx: Gradient of the loss with respect to x
-    """
-    N = x.shape[0]
-    correct_class_scores = x[np.arange(N), y]
-    margins = np.maximum(0, x - correct_class_scores[:, np.newaxis] + 1.0)
-    margins[np.arange(N), y] = 0
-    loss = np.sum(margins) / N
-    num_pos = np.sum(margins > 0, axis=1)
-    dx = np.zeros_like(x)
-    dx[margins > 0] = 1
-    dx[np.arange(N), y] -= num_pos
-    dx /= N
-    return loss, dx
-
-
-def softmax_loss(x, y):
-    """
-    Computes the loss and gradient for softmax classification.
-
-    Inputs:
-    - x: Input data, of shape (N, C) where x[i, j] is the score for the jth class
-      for the ith input.
-    - y: Vector of labels, of shape (N,) where y[i] is the label for x[i] and
-      0 <= y[i] < C
-
-    Returns a tuple of:
-    - loss: Scalar giving the loss
-    - dx: Gradient of the loss with respect to x
-    """
-    probs = np.exp(x - np.max(x, axis=1, keepdims=True))
-    probs /= np.sum(probs, axis=1, keepdims=True)
-    N = x.shape[0]
-    loss = -np.sum(np.log(probs[np.arange(N), y])) / N
-    dx = probs.copy()
-    dx[np.arange(N), y] -= 1
-    dx /= N
-    return loss, dx
