--- cs231n/optim.py
@@ -1,83 +1,0 @@
-import numpy as np
-
-"""
-This file implements various first-order update rules that are commonly used for
-training neural networks. Each update rule accepts current weights and the
-gradient of the loss with respect to those weights and produces the next set of
-weights. Each update rule has the same interface:
-
-def update(w, dw, config=None):
-
-Inputs:
-  - w: A numpy array giving the current weights.
-  - dw: A numpy array of the same shape as w giving the gradient of the
-    loss with respect to w.
-  - config: A dictionary containing hyperparameter values such as learning rate,
-    momentum, etc. If the update rule requires caching values over many
-    iterations, then config will also hold these cached values.
-
-Returns:
-  - next_w: The next point after the update.
-  - config: The config dictionary to be passed to the next iteration of the
-    update rule.
-
-NOTE: For most update rules, the default learning rate will probably not perform
-well; however the default values of the other hyperparameters should work well
-for a variety of different problems.
-
-For efficiency, update rules may perform in-place updates, mutating w and
-setting next_w equal to w.
-"""
-
-
-def sgd(w, dw, config=None):
-    """
-    Performs vanilla stochastic gradient descent.
-
-    config format:
-    - learning_rate: Scalar learning rate.
-    """
-    if config is None: config = {}
-    config.setdefault('learning_rate', 1e-2)
-
-    w -= config['learning_rate'] * dw
-    return w, config
-
-
-def adam(x, dx, config=None):
-    """
-    Uses the Adam update rule, which incorporates moving averages of both the
-    gradient and its square and a bias correction term.
-
-    config format:
-    - learning_rate: Scalar learning rate.
-    - beta1: Decay rate for moving average of first moment of gradient.
-    - beta2: Decay rate for moving average of second moment of gradient.
-    - epsilon: Small scalar used for smoothing to avoid dividing by zero.
-    - m: Moving average of gradient.
-    - v: Moving average of squared gradient.
-    - t: Iteration number.
-    """
-    if config is None: config = {}
-    config.setdefault('learning_rate', 1e-3)
-    config.setdefault('beta1', 0.9)
-    config.setdefault('beta2', 0.999)
-    config.setdefault('epsilon', 1e-8)
-    config.setdefault('m', np.zeros_like(x))
-    config.setdefault('v', np.zeros_like(x))
-    config.setdefault('t', 0)
-
-    next_x = None
-    beta1, beta2, eps = config['beta1'], config['beta2'], config['epsilon']
-    t, m, v = config['t'], config['m'], config['v']
-    m = beta1 * m + (1 - beta1) * dx
-    v = beta2 * v + (1 - beta2) * (dx * dx)
-    t += 1
-    alpha = config['learning_rate'] * np.sqrt(1 - beta2 ** t) / (1 - beta1 ** t)
-    x -= alpha * (m / (np.sqrt(v) + eps))
-    config['t'] = t
-    config['m'] = m
-    config['v'] = v
-    next_x = x
-
-    return next_x, config
