--- cs231n/rnn_layers.py
@@ -1,563 +1,0 @@
-from __future__ import print_function, division
-from builtins import range
-import numpy as np
-
-
-"""
-This file defines layer types that are commonly used for recurrent neural
-networks.
-"""
-
-
-def rnn_step_forward(x, prev_h, Wx, Wh, b):
-    """
-    Run the forward pass for a single timestep of a vanilla RNN that uses a tanh
-    activation function.
-
-    The input data has dimension D, the hidden state has dimension H, and we use
-    a minibatch size of N.
-
-    Inputs:
-    - x: Input data for this timestep, of shape (N, D).
-    - prev_h: Hidden state from previous timestep, of shape (N, H)
-    - Wx: Weight matrix for input-to-hidden connections, of shape (D, H)
-    - Wh: Weight matrix for hidden-to-hidden connections, of shape (H, H)
-    - b: Biases of shape (H,)
-
-    Returns a tuple of:
-    - next_h: Next hidden state, of shape (N, H)
-    - cache: Tuple of values needed for the backward pass.
-    """
-    next_h, cache = None, None
-    c = -100
-    affined_wh = np.matmul(prev_h,Wh)
-    cache_affined_wh = (prev_h,Wh)
-    affined_wx = np.matmul(x,Wx)
-    cache_affined_wx = (x,Wx)
-
-    sum_affined = affined_wh + affined_wx + b
-
-    jisu = sum_affined*(-2) + c
-    exp_aff = np.exp(jisu)
-    tanh_aff = (np.exp(c)-exp_aff)/(exp_aff+np.exp(c))
-
-    next_h = tanh_aff
-    cache = (affined_wh,cache_affined_wh,affined_wx,cache_affined_wx,sum_affined,exp_aff,tanh_aff)
-    ##############################################################################
-    # Implement a single forward step for the vanilla RNN. Store the next  #
-    # hidden state and any values you need for the backward pass in the next_h   #
-    # and cache variables respectively.                                          #
-    ##############################################################################
-
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-    return next_h, cache
-
-
-def rnn_step_backward(dnext_h, cache):
-    """
-    Backward pass for a single timestep of a vanilla RNN.
-
-    Inputs:
-    - dnext_h: Gradient of loss with respect to next hidden state
-    - cache: Cache object from the forward pass
-
-    Returns a tuple of:
-    - dx: Gradients of input data, of shape (N, D)
-    - dprev_h: Gradients of previous hidden state, of shape (N, H)
-    - dWx: Gradients of input-to-hidden weights, of shape (D, H)
-    - dWh: Gradients of hidden-to-hidden weights, of shape (H, H)
-    - db: Gradients of bias vector, of shape (H,)
-    """
-    dx, dprev_h, dWx, dWh, db = None, None, None, None, None
-
-    affined_wh, cache_affined_wh, affined_wx, cache_affined_wx, sum_affined, exp_aff,tan_aff = cache
-
-    prev_h,Wh = cache_affined_wh
-    x,Wx = cache_affined_wx
-    #dexp = (-2 * (1+exp_aff)**(-2)) * dnext_h
-
-    dsum = (1- tan_aff**2 ) * dnext_h
-
-    db = np.sum(dsum,axis=0)
-
-    #daffined_wh = dsum
-    dprev_h = np.matmul(dsum,Wh.T)
-    dWh = np.matmul(prev_h.T,dsum)
-
-    #daffined_wx = dsum
-    dx = np.matmul(dsum,Wx.T)
-    dWx = np.matmul(x.T,dsum)
-
-
-
-
-
-    ##############################################################################
-    #  Implement the backward pass for a single step of a vanilla RNN.      #
-    #                                                                            #
-    # HINT: For the tanh function, you can compute the local derivative in terms #
-    # of the output value from tanh.                                             #
-    ##############################################################################
-
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-    return dx, dprev_h, dWx, dWh, db
-
-
-def rnn_forward(x, h0, Wx, Wh, b):
-    """
-    Run a vanilla RNN forward on an entire sequence of data. We assume an input
-    sequence composed of T vectors, each of dimension D. The RNN uses a hidden
-    size of H, and we work over a minibatch containing N sequences. After running
-    the RNN forward, we return the hidden states for all timesteps.
-
-    Inputs:
-    - x: Input data for the entire timeseries, of shape (N, T, D).
-    - h0: Initial hidden state, of shape (N, H)
-    - Wx: Weight matrix for input-to-hidden connections, of shape (D, H)
-    - Wh: Weight matrix for hidden-to-hidden connections, of shape (H, H)
-    - b: Biases of shape (H,)
-
-    Returns a tuple of:
-    - h: Hidden states for the entire timeseries, of shape (N, T, H).
-    - cache: Values needed in the backward pass
-    """
-    h, cache = None,[]
-    N,T,D = x.shape
-    x_li = np.transpose(x,axes=(1,0,2))
-    h_li = []
-    tmp_h = h0
-    for i in range(T):
-        tmp_h,tmp_cache = rnn_step_forward(x_li[i],Wh=Wh,Wx=Wx,prev_h=tmp_h,b=b)
-        h_li.append(tmp_h)
-        cache.append(tmp_cache)
-    ##############################################################################
-    # Implement forward pass for a vanilla RNN running on a sequence of    #
-    # input data. You should use the rnn_step_forward function that you defined  #
-    # above. You can use a for loop to help compute the forward pass.            #
-    ##############################################################################
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-    h = np.array(h_li).transpose((1,0,2))
-    return h, cache
-
-
-def rnn_backward(dh, cache):
-    """
-    Compute the backward pass for a vanilla RNN over an entire sequence of data.
-
-    Inputs:
-    - dh: Upstream gradients of all hidden states, of shape (N, T, H)
-
-    Returns a tuple of:
-    - dx: Gradient of inputs, of shape (N, T, D)
-    - dh0: Gradient of initial hidden state, of shape (N, H)
-    - dWx: Gradient of input-to-hidden weights, of shape (D, H)
-    - dWh: Gradient of hidden-to-hidden weights, of shape (H, H)
-    - db: Gradient of biases, of shape (H,)
-    """
-    dx, dh0, dWx, dWh, db = None, None, None, None, None
-    N,T,H = dh.shape
-    D,H = cache[-1][3][1].shape
-
-    dh_li = np.transpose(dh,axes=(1,0,2))
-    dh0 = np.ones_like(dh_li[0])
-    dx_li = np.zeros((T,N,D))
-    dWh = np.zeros((H,H))
-    dWx = np.zeros((D,H))
-    db = np.zeros((H,))
-
-    tmp_dprev_h = np.zeros_like(dh_li[-1])
-    for i in range(1,T+1):
-        tmp_dprev_h += dh_li[-i]
-        tmp_dx, tmp_dprev_h, tmp_dWx, tmp_dWh, tmp_db = rnn_step_backward(dnext_h=tmp_dprev_h,cache= cache[-i])
-        dx_li[T-i,:,:] = tmp_dx
-        dWh += tmp_dWh
-        dWx += tmp_dWx
-        dh0 = tmp_dprev_h
-        db+= tmp_db
-    dx=np.transpose(dx_li,axes=(1,0,2))
-   
-
-    ##############################################################################
-    # Implement the backward pass for a vanilla RNN running an entire      #
-    # sequence of data. You should use the rnn_step_backward function that you   #
-    # defined above. You can use a for loop to help compute the backward pass.   #
-    # key : rnn_step_backward에 집어넣을 dh는 메소드의 인풋으로 받은 dh 뿐만 아니라, 이전단계의 flow #
-    # (윗단계prop)에서 획득한 dh(tmp_prev_h) 를 더하여 넣어주여야 한다.                     #
-    ##############################################################################
-
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-    return dx, dh0, dWx, dWh, db
-
-def word_embedding_forward(x, W):
-    """
-    Forward pass for word embeddings. We operate on minibatches of size N where
-    each sequence has length T. We assume a vocabulary of V words, assigning each
-    to a vector of dimension D.
-
-    Inputs:
-    - x: Integer array of shape (N, T) giving indices of words. Each element idx
-      of x muxt be in the range 0 <= idx < V.
-    - W: Weight matrix of shape (V, D) giving word vectors for all words.
-
-    Returns a tuple of:
-    - out: Array of shape (N, T, D) giving word vectors for all input words.
-    - cache: Values needed for the backward pass
-    """
-    out, cache = None, None
-    N,T = x.shape
-    V,D = W.shape
-    out = np.zeros((N,T,D))
-
-    out = W[x]
-    cache = (x,W)
-    ##############################################################################
-    # Implement the forward pass for word embeddings.                      #
-    #                                                                            #
-    # HINT: This can be done in one line using NumPy's array indexing.           #
-    ##############################################################################
-
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-    return out, cache
-
-def word_embedding_backward(dout, cache):
-    """
-    Backward pass for word embeddings. We cannot back-propagate into the words
-    since they are integers, so we only return gradient for the word embedding
-    matrix.
-
-    HINT: Look up the function np.add.at
-
-    Inputs:
-    - dout: Upstream gradients of shape (N, T, D)
-    - cache: Values from the forward pass
-
-    Returns:
-    - dW: Gradient of word embedding matrix, of shape (V, D).
-    """
-    x, W = cache
-    N,T,D = dout.shape
-    dW = np.zeros_like(W)
-    np.add.at(dW,x.reshape(N*T),dout.reshape(N*T,D))
-    ##############################################################################
-    # Implement the backward pass for word embeddings.                     #
-    #                                                                            #
-    # Note that Words can appear more than once in a sequence.                   #
-    # HINT: Look up the function np.add.at                                       #
-    ##############################################################################
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-    return dW
-
-def sigmoid(x):
-    """
-    A numerically stable version of the logistic sigmoid function.
-    """
-    pos_mask = (x >= 0)
-    neg_mask = (x < 0)
-    z = np.zeros_like(x)
-    z[pos_mask] = np.exp(-x[pos_mask])
-    z[neg_mask] = np.exp(x[neg_mask])
-    top = np.ones_like(x)
-    top[neg_mask] = z[neg_mask]
-    return top / (1 + z)
-
-
-def lstm_step_forward(x, prev_h, prev_c, Wx, Wh, b):
-    """
-    Forward pass for a single timestep of an LSTM.
-
-    The input data has dimension D, the hidden state has dimension H, and we use
-    a minibatch size of N.
-
-    Inputs:
-    - x: Input data, of shape (N, D)
-    - prev_h: Previous hidden state, of shape (N, H)
-    - prev_c: previous cell state, of shape (N, H)
-    - Wx: Input-to-hidden weights, of shape (D, 4H)
-    - Wh: Hidden-to-hidden weights, of shape (H, 4H)
-    - b: Biases, of shape (4H,)
-
-    Returns a tuple of:
-    - next_h: Next hidden state, of shape (N, H)
-    - next_c: Next cell state, of shape (N, H)
-    - cache: Tuple of values needed for backward pass.
-    """
-    next_h, next_c, cache = None, None, None
-    N,H = prev_h.shape
-    affined = np.matmul(x,Wx) + np.matmul(prev_h,Wh) + b
-    i = sigmoid(affined[:,0:H])
-    f = sigmoid(affined[:,H:2*H])
-    o = sigmoid(affined[:,2*H:3*H])
-    g = np.tanh(affined[:,3*H:4*H])
-
-    next_c =f*prev_c + i * g
-    tanhed_ct = np.tanh(next_c)
-    next_h = o *tanhed_ct
-
-    cache = (Wx,Wh,x,prev_h,prev_c,affined,i,f,o,g,next_c,next_h)
-
-
-    #############################################################################
-    # Implement the forward pass for a single timestep of an LSTM.        #
-    # You may want to use the numerically stable sigmoid implementation above.  #
-    #############################################################################
-    pass
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-
-    return next_h, next_c, cache
-
-
-def lstm_step_backward(dnext_h, dnext_c, cache):
-    """
-    Backward pass for a single timestep of an LSTM.
-
-    Inputs:
-    - dnext_h: Gradients of next hidden state, of shape (N, H)
-    - dnext_c: Gradients of next cell state, of shape (N, H)
-    - cache: Values from the forward pass
-
-    Returns a tuple of:
-    - dx: Gradient of input data, of shape (N, D)
-    - dprev_h: Gradient of previous hidden state, of shape (N, H)
-    - dprev_c: Gradient of previous cell state, of shape (N, H)
-    - dWx: Gradient of input-to-hidden weights, of shape (D, 4H)
-    - dWh: Gradient of hidden-to-hidden weights, of shape (H, 4H)
-    - db: Gradient of biases, of shape (4H,)
-    """
-    dx, dh, dc, dWx, dWh, db = None, None, None, None, None, None
-
-    Wx,Wh,x,prev_h,prev_c, affined, i, f, o, g, next_c, next_h =cache
-
-    N,H = dnext_h.shape
-    dtanh_next_c = o * dnext_h
-    do = np.tanh(next_c) * dnext_h
-
-    dnext_c += (1- (np.tanh(next_c) ** 2)) * dtanh_next_c
-    df = dnext_c * (prev_c)
-    dprev_c = dnext_c * f
-    di = dnext_c * g
-    dg = dnext_c * i
-
-    sums = np.zeros(shape=(N,4*H))
-    sums[:,0:H] = di * (1-i) * (i)
-    sums[:,H:2*H] = df * (1 - f) * (f)
-    sums[:,2*H:3*H] = do * (1 - o) *(o)
-    sums[:,3*H:4*H] = dg * (1 - (g **2))
-
-    db = np.sum(sums,axis=0)
-    dWh = np.matmul(prev_h.T, sums)
-    dWx = np.matmul(x.T,sums)
-    dprev_h = np.matmul(sums,Wh.T)
-    dx = np.matmul(sums,Wx.T)
-    #############################################################################
-    # Implement the backward pass for a single timestep of an LSTM.       #
-    #                                                                           #
-    # HINT: For sigmoid and tanh you can compute local derivatives in terms of  #
-    # the output value from the nonlinearity.                                   #
-    #############################################################################
-
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-
-
-    return dx, dprev_h, dprev_c, dWx, dWh, db
-
-
-def lstm_forward(x, h0, Wx, Wh, b):
-    """
-    Forward pass for an LSTM over an entire sequence of data. We assume an input
-    sequence composed of T vectors, each of dimension D. The LSTM uses a hidden
-    size of H, and we work over a minibatch containing N sequences. After running
-    the LSTM forward, we return the hidden states for all timesteps.
-
-    Note that the initial cell state is passed as input, but the initial cell
-    state is set to zero. Also note that the cell state is not returned; it is
-    an internal variable to the LSTM and is not accessed from outside.
-
-    Inputs:
-    - x: Input data of shape (N, T, D)
-    - h0: Initial hidden state of shape (N, H)
-    - Wx: Weights for input-to-hidden connections, of shape (D, 4H)
-    - Wh: Weights for hidden-to-hidden connections, of shape (H, 4H)
-    - b: Biases of shape (4H,)
-
-    Returns a tuple of:
-    - h: Hidden states for all timesteps of all sequences, of shape (N, T, H)
-    - cache: Values needed for the backward pass.
-    """
-    h, cache = None, []
-    N,T,D = x.shape
-    H, _ = Wh.shape
-    h_li=[]
-    x_li = np.transpose(x,axes=(1,0,2))
-    init_c = np.zeros((N,H))
-    init_c.astype(dtype=float)
-    next_h = h0
-    next_c = init_c
-    for i in range(T):
-        next_h,next_c,tmp_cache = lstm_step_forward(x_li[i],next_h,next_c,Wx,Wh,b)
-        h_li.append(next_h)
-        cache.append(tmp_cache)
-
-    h = np.array(h_li).transpose((1,0,2))
-    #############################################################################
-    # Implement the forward pass for an LSTM over an entire timeseries.   #
-    # You should use the lstm_step_forward function that you just defined.      #
-    #############################################################################
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-
-    return h, cache
-
-
-def lstm_backward(dh, cache):
-    """
-    Backward pass for an LSTM over an entire sequence of data.]
-
-    Inputs:
-    - dh: Upstream gradients of hidden states, of shape (N, T, H)
-    - cache: Values from the forward pass
-
-    Returns a tuple of:
-    - dx: Gradient of input data of shape (N, T, D)
-    - dh0: Gradient of initial hidden state of shape (N, H)
-    - dWx: Gradient of input-to-hidden weight matrix of shape (D, 4H)
-    - dWh: Gradient of hidden-to-hidden weight matrix of shape (H, 4H)
-    - db: Gradient of biases, of shape (4H,)
-    """
-
-    N,T,H = dh.shape
-    _,D = cache[-1][2].shape
-    dh_li = np.transpose(dh,axes=(1,0,2))
-    dprev_c = np.zeros(shape=(N,H))
-    dx_li = []
-    dprev_h = 0
-    dx, dh0, dWx, dWh, db = None, None, np.zeros((D,4*H)),np.zeros((H,4*H)), np.zeros((4*H,))
-    for i in range(1,T+1):
-        tmp_dx,dprev_h,dprev_c,tmp_dWx,tmp_dWh,tmp_db = lstm_step_backward(dh_li[-i] + dprev_h,dprev_c,cache=cache[-i])
-        dx_li.append(tmp_dx)
-        dWx += tmp_dWx
-        dWh += tmp_dWh
-        db += tmp_db
-    dx_li.reverse()
-    dx = np.array(dx_li).transpose((1,0,2))
-    dh0 = dprev_h
-    #############################################################################
-    # Implement the backward pass for an LSTM over an entire timeseries.  #
-    # You should use the lstm_step_backward function that you just defined.     #
-    #############################################################################
-    ##############################################################################
-    #                               END OF YOUR CODE                             #
-    ##############################################################################
-
-    return dx, dh0, dWx, dWh, db
-
-
-def temporal_affine_forward(x, w, b):
-    """
-    Forward pass for a temporal affine layer. The input is a set of D-dimensional
-    vectors arranged into a minibatch of N timeseries, each of length T. We use
-    an affine function to transform each of those vectors into a new vector of
-    dimension M.
-
-    Inputs:
-    - x: Input data of shape (N, T, D)
-    - w: Weights of shape (D, M)
-    - b: Biases of shape (M,)
-
-    Returns a tuple of:
-    - out: Output data of shape (N, T, M)
-    - cache: Values needed for the backward pass
-    """
-    N, T, D = x.shape
-    M = b.shape[0]
-    out = x.reshape(N * T, D).dot(w).reshape(N, T, M) + b
-    cache = x, w, b, out
-    return out, cache
-
-
-def temporal_affine_backward(dout, cache):
-    """
-    Backward pass for temporal affine layer.
-
-    Input:
-    - dout: Upstream gradients of shape (N, T, M)
-    - cache: Values from forward pass
-
-    Returns a tuple of:
-    - dx: Gradient of input, of shape (N, T, D)
-    - dw: Gradient of weights, of shape (D, M)
-    - db: Gradient of biases, of shape (M,)
-    """
-    x, w, b, out = cache
-    N, T, D = x.shape
-    M = b.shape[0]
-
-    dx = dout.reshape(N * T, M).dot(w.T).reshape(N, T, D)
-    dw = dout.reshape(N * T, M).T.dot(x.reshape(N * T, D)).T
-    db = dout.sum(axis=(0, 1))
-
-    return dx, dw, db
-
-
-def temporal_softmax_loss(x, y, mask, verbose=False):
-    """
-    A temporal version of softmax loss for use in RNNs. We assume that we are
-    making predictions over a vocabulary of size V for each timestep of a
-    timeseries of length T, over a minibatch of size N. The input x gives scores
-    for all vocabulary elements at all timesteps, and y gives the indices of the
-    ground-truth element at each timestep. We use a cross-entropy loss at each
-    timestep, summing the loss over all timesteps and averaging across the
-    minibatch.
-
-    As an additional complication, we may want to ignore the model output at some
-    timesteps, since sequences of different length may have been combined into a
-    minibatch and padded with NULL tokens. The optional mask argument tells us
-    which elements should contribute to the loss.
-
-    Inputs:
-    - x: Input scores, of shape (N, T, V)
-    - y: Ground-truth indices, of shape (N, T) where each element is in the range
-         0 <= y[i, t] < V
-    - mask: Boolean array of shape (N, T) where mask[i, t] tells whether or not
-      the scores at x[i, t] should contribute to the loss.
-
-    Returns a tuple of:
-    - loss: Scalar giving loss
-    - dx: Gradient of loss with respect to scores x.
-    """
-
-    N, T, V = x.shape
-
-    x_flat = x.reshape(N * T, V)
-    y_flat = y.reshape(N * T)
-    mask_flat = mask.reshape(N * T)
-
-    probs = np.exp(x_flat - np.max(x_flat, axis=1, keepdims=True))
-    probs /= np.sum(probs, axis=1, keepdims=True)
-    loss = -np.sum(mask_flat * np.log(probs[np.arange(N * T), y_flat])) / N
-    dx_flat = probs.copy()
-    dx_flat[np.arange(N * T), y_flat] -= 1
-    dx_flat /= N
-    dx_flat *= mask_flat[:, None]
-
-    if verbose: print('dx_flat: ', dx_flat.shape)
-
-    dx = dx_flat.reshape(N, T, V)
-
-    return loss, dx
