--- start_ipython_osx.sh
@@ -1,4 +1,0 @@
-# Assume the virtualenv is called .env
-
-cp frameworkpython .env/bin
-.env/bin/frameworkpython -m IPython notebook
