+++ style_transfer/datasets/get_assignment3_data.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+./get_coco_captioning.sh
+./get_squeezenet_tf.sh
+./get_imagenet_val.sh
+
