+++ style_transfer/datasets/get_coco_captioning.sh
@@ -1,0 +1,3 @@
+wget "http://cs231n.stanford.edu/coco_captioning.zip"
+unzip coco_captioning.zip
+rm coco_captioning.zip
