+++ style_transfer/datasets/get_imagenet_val.sh
@@ -1,0 +1,1 @@
+wget http://cs231n.stanford.edu/imagenet_val_25.npz
