+++ style_transfer/datasets/get_squeezenet_tf.sh
@@ -1,0 +1,3 @@
+wget "http://cs231n.stanford.edu/squeezenet_tf.zip"
+unzip squeezenet_tf.zip
+rm squeezenet_tf.zip
