--- where_are_my_drivers.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-echo "Checking for CUDA and installing."
-# Check for CUDA and try to install.
-if ! dpkg-query -W cuda; then
-  # The 16.04 installer works with 16.10.
-  curl -O http://developer.download.nvidia.com/compute/cuda/repos/ubuntu1604/x86_64/cuda-repo-ubuntu1604_8.0.61-1_amd64.deb
-  dpkg -i ./cuda-repo-ubuntu1604_8.0.61-1_amd64.deb
-  apt-get update
-  apt-get install cuda -y
-fi
